+++ src/Module/Api/Method/GetExternalMetadataMethod.php
@@ -25,12 +25,14 @@ declare(strict_types=0);
 
 namespace Ampache\Module\Api\Method;
 
+use Ampache\Config\AmpConfig;
 use Ampache\Module\Api\Api;
 use Ampache\Module\Api\Exception\ErrorCodeEnum;
 use Ampache\Module\Api\Xml_Data;
 use Ampache\Module\System\Plugin\PluginTypeEnum;
 use Ampache\Repository\Model\Album;
 use Ampache\Repository\Model\Artist;
+use Ampache\Repository\Model\Label;
 use Ampache\Repository\Model\Plugin;
 use Ampache\Repository\Model\Song;
 use Ampache\Repository\Model\User;
@@ -49,7 +51,7 @@ final class GetExternalMetadataMethod
      *
      * Return External plugin metadata searching by object id and type
      *
-     * type   = (string) 'song', 'artist', 'album'
+     * type   = (string) 'song', 'artist', 'album', 'label'
      * filter = (integer) album id, artist id or song id
      */
     public static function get_external_metadata(array $input, User $user): bool
@@ -60,11 +62,18 @@ final class GetExternalMetadataMethod
         $type      = (string) $input['type'];
         $object_id = (int) $input['filter'];
         // confirm the correct data
-        if (!in_array(strtolower($type), ['song', 'album', 'artist'])) {
+        if (!in_array(strtolower($type), ['song', 'album', 'artist', 'label'])) {
             Api::error(sprintf('Bad Request: %s', $type), ErrorCodeEnum::BAD_REQUEST, self::ACTION, 'type', $input['api_format']);
 
             return false;
         }
+
+        if ($type == 'label' && !AmpConfig::get('label')) {
+            Api::error('Enable: label', ErrorCodeEnum::ACCESS_DENIED, self::ACTION, 'system', $input['api_format']);
+
+            return false;
+        }
+
         switch ($type) {
             case 'song':
                 $libitem = new Song($object_id);
@@ -87,6 +96,13 @@ final class GetExternalMetadataMethod
                 $data    = [
                     'artist' => $libitem->get_fullname(),
                     'mb_artistid' => $libitem->mbid,
+                ];
+                break;
+            case 'label':
+                $libitem = new Label($object_id);
+                $data    = [
+                    'label' => $libitem->get_fullname(),
+                    'mb_labelid' => $libitem->mbid,
                 ];
         }
         if (!isset($data) || !isset($libitem) || $libitem->isNew()) {