+++ src/Module/Util/Rss/Type/NowPlayingFeed.php
@@ -62,17 +62,17 @@ final readonly class NowPlayingFeed extends AbstractGenericRssFeed
         $data    = Stream::get_now_playing();
         $element = array_shift($data);
 
-        /** @var Song|Video $media */
-        $media   = $element['media'] ?? null;
-        if (!$media) {
+        /** @var Song|Video|null $media */
+        $media = $element['media'] ?? null;
+        if ($media === null) {
             return null;
         }
 
-        $type    = ($media instanceof Video)
+        $type = ($media instanceof Video)
             ? 'video'
             : 'song';
 
-        return (string)Art::url($media->getId(), $type, null, 2) ?? null;
+        return (string)Art::url($media->getId(), $type, null, 2);
     }
 
     protected function getItems(): Generator
@@ -112,7 +112,12 @@ final readonly class NowPlayingFeed extends AbstractGenericRssFeed
                 'description' => str_replace('<p>Artist: </p><p>Album: </p>', '', $description),
                 'comments' => $client->get_link(),
                 'pubDate' => date("r", (int)$element['expire']),
-                'guid' => $element['expire'] . '-' . $client->getId() . '-' . $media->getId(),
+                'guid' => (isset($media->mbid))
+                    ? 'https://musicbrainz.org/recording/' . $media->mbid
+                    : $element['expire'] . '-' . $client->getId() . '-' . $media->getId(),
+                'isPermaLink' => (isset($media->mbid))
+                    ? 'true'
+                    : 'false',
             ];
         }
     }